add a way to fail for shared lib
in regular c extensions.

so, there's a way to fail for ffi style loading.
so, there's a way to fail for C ext shared libs.

those can be used for testing.
make some bundle plugin (read: use the bundler library, but just make a
standalone library/executable which bundler.which will exec).
make a pull request out of that at some point.